Add frame saving on keypress
To be added is a 'raycast' conversion for clicking on the grid to toggle the seed
on or off (alive/dead). As well as a way to randomly seed the grid, and display the
population number onscreen in an unobtrusive way. Finally, a possible branch/new project
will be pursued wherein the user creates a seed, runs the simulation and then at
any point can have it converted into a moving Rorscach. Tashiv suggested combining
the Conway sim and a moving Rorscach at the same time and fading in/out inkblots
when they are alive/dead.